[ProjectPage] Deprecate `ContributionCard`
