+++ assets/javascripts/kitten/helpers/utils/deprecated.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react'
+
+export class Deprecated extends Component {
+  componentWillMount() {
+    if (process.env.NODE_ENV === 'development') {
+      const componentName = React.Children.only(this.props.children)._owner.type
+        .name
+
+      console.warn(
+        `Warning: ${componentName} is deprecated.`,
+        this.props.warningMessasge,
+      )
+    }
+  }
+
+  render() {
+    return this.props.children
+  }
+}
